Rating system implemented

+++ db/migrate/20160606045920_create_reviews.rb
@@ -2,7 +2,7 @@ class CreateReviews < ActiveRecord::Migration
   def change
     create_table :reviews do |t|
       t.text :content
-      t.integer :rating
+      t.float :rating
       t.integer :user_id
       t.integer :movie_id
